centro practica arreglado

+++ protected/controllers/ConfiguracionpracticaController.php
@@ -136,11 +136,24 @@ class ConfiguracionpracticaController extends Controller
 	 */
 	public function actionDelete($id)
 	{
-		$this->loadModel($id)->delete();
-
-		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-		if(!isset($_GET['ajax']))
-			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+        $table = "estudiante";
+		$codTable = "ConfiguracionPractica_NombrePractica";
+		
+		$exist = contains($table,$codTable,$id);
+		
+		if($exist == 0){
+            $this->loadModel($id)->delete();
+            
+            // if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
+            if(!isset($_GET['ajax']))
+                $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+        }else{
+            Yii::app()->user->setFlash('message',"<div id='errorMessage' class='flash-error'><p><strong>¡No es posible eliminar!</strong></p><ul><li>Hay estudiantes asociados a esta práctica.</li></ul></div>");
+			//$this->refresh();
+			$this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array("view&id=".$id.""));
+        }
+        
+		
 	}
 
 	/**
